# src/api/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.CoinDeletedResponse:
    properties:
      message:
        example: Coin deleted
        type: string
    type: object
  handlers.CoinListResponse:
    properties:
      coins:
        items:
          $ref: '#/definitions/models.Coin'
        type: array
      limit:
        example: 50
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 42
        type: integer
    type: object
...
  handlers.ErrorResponse:
    properties:
      error:
        example: Something went wrong
        type: string
    type: object
...
  handlers.PurchaseRequest:
    properties:
      purchaseDate:
        type: string
      purchaseLocation:
        type: string
      purchasePrice:
        type: number
    type: object
...
  models.Coin:
    properties:
      aiAnalysis:
        type: string
      category:
        $ref: '#/definitions/models.Category'
      createdAt:
        type: string
      currentValue:
        type: number
      currentValueUpdatedAt:
        type: string
      denomination:
        maxLength: 200
        type: string
      diameterMm:
        type: number
      era:
        allOf:
        - $ref: '#/definitions/models.Era'
        enum:
        - ancient
        - medieval
        - modern
      grade:
        maxLength: 100
        type: string
      id:
        type: integer
      images:
        items:
          $ref: '#/definitions/models.CoinImage'
        type: array
      isPrivate:
        type: boolean
      isSold:
        type: boolean
      isWishlist:
        type: boolean
      listingCheckReason:
        type: string
      listingCheckedAt:
        type: string
      listingStatus:
        type: string
      material:
        $ref: '#/definitions/models.Material'
      mint:
        maxLength: 200
        type: string
      name:
        maxLength: 200
        type: string
      notes:
        maxLength: 5000
        type: string
      obverseAnalysis:
        type: string
      obverseDescription:
        maxLength: 2000
        type: string
      obverseInscription:
        maxLength: 1000
        type: string
      purchaseDate:
        type: string
      purchaseLocation:
        maxLength: 500
        type: string
      purchasePrice:
        type: number
      rarityRating:
        maxLength: 100
        type: string
      referenceText:
        maxLength: 2000
        type: string
      referenceUrl:
        maxLength: 2000
        type: string
      references:
        items:
          $ref: '#/definitions/models.CoinReference'
        type: array
      reverseAnalysis:
        type: string
      reverseDescription:
        maxLength: 2000
        type: string
      reverseInscription:
        maxLength: 1000
        type: string
      ruler:
        maxLength: 200
        type: string
      soldDate:
        type: string
      soldPrice:
        type: number
      soldTo:
        type: string
      storageLocation:
        $ref: '#/definitions/models.StorageLocation'
      storageLocationId:
        type: integer
      tags:
        items:
          $ref: '#/definitions/models.Tag'
        type: array
      updatedAt:
        type: string
      userId:
        type: integer
      weightGrams:
        type: number
    type: object
...
  models.CoinValueHistory:
    properties:
      coinId:
        type: integer
      confidence:
        type: string
      id:
        type: integer
      recordedAt:
        type: string
      userId:
        type: integer
      value:
        type: number
    type: object
...
info:
  contact: {}
  description: REST API for managing an ancient coin collection. Supports coin CRUD,
    image uploads, AI-powered analysis via Ollama, user management, and admin features.
  title: Ancient Coins API
  version: "1.0"
paths:
...
  /coins:
    get:
      description: Returns a paginated, filterable list of coins belonging to the
        authenticated user.
      parameters:
      - description: Filter by category (Roman, Greek, Byzantine, Modern, Other)
        in: query
        name: category
        type: string
      - description: Filter by era (ancient, medieval, modern)
        in: query
        name: era
        type: string
      - description: Search across name, denomination, ruler, era, mint, inscriptions,
          notes
        in: query
        name: search
        type: string
      - description: Filter by wishlist status
        enum:
        - "true"
        - "false"
        in: query
        name: wishlist
        type: string
      - description: Filter by sold status
        enum:
        - "true"
        - "false"
        in: query
        name: sold
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - default: updated_at
        description: Sort field
        enum:
        - created_at
        - updated_at
        - current_value
        in: query
        name: sort
        type: string
      - default: desc
        description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CoinListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List coins
      tags:
      - Coins
    post:
      consumes:
      - application/json
      description: Creates a new coin record for the authenticated user.
      parameters:
      - description: Coin data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Coin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Coin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a coin
      tags:
      - Coins
  /coins/{id}:
    delete:
      description: Deletes a coin and all associated image records. Only the coin
        owner can delete it.
      parameters:
      - description: Coin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.CoinDeletedResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a coin
      tags:
      - Coins
    get:
      description: Returns a single coin with its images, owned by the authenticated
        user.
      parameters:
      - description: Coin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Coin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get a coin
      tags:
      - Coins
    put:
      consumes:
      - application/json
      description: Updates an existing coin record. Only the coin owner can update
        it.
      parameters:
      - description: Coin ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated coin data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Coin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Coin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a coin
      tags:
      - Coins
...
  /coins/{id}/purchase:
    post:
      consumes:
      - application/json
      description: Sets isWishlist to false, moving the coin from wishlist to collection.
        Optionally accepts purchase details.
      parameters:
      - description: Coin ID
        in: path
        name: id
        required: true
        type: integer
      - description: Optional purchase details
        in: body
        name: body
        schema:
          $ref: '#/definitions/handlers.PurchaseRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Coin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark coin as purchased
      tags:
      - Coins
...
  /coins/{id}/sell:
    post:
      consumes:
      - application/json
      description: Sets isSold to true with a sold price and date, moving the coin
        from collection to sold gallery.
      parameters:
      - description: Coin ID
        in: path
        name: id
        required: true
        type: integer
      - description: Sale details
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Coin'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Mark coin as sold
      tags:
      - Coins
  /coins/{id}/value-history:
    get:
      description: Returns value tracking entries for a coin, ordered oldest first.
      parameters:
      - description: Coin ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CoinValueHistory'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get coin value history
      tags:
      - Coins
...
  /stats:
    get:
      description: Returns counts by category/material, total/average values, and
        wishlist count.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.StatsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get collection statistics
      tags:
      - Coins
  /stats/distribution:
    get:
      description: Returns era × category counts for the heat map visualization.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/repository.DistributionCell'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get collection distribution
      tags:
      - Coins
...
  /suggestions:
    get:
      description: Returns distinct values for the specified field, optionally filtered
        by a search query. Limited to 20 results.
      parameters:
      - description: Field to get suggestions for
        enum:
        - name
        - denomination
        - purchaseLocation
        in: query
        name: field
        required: true
        type: string
      - description: Search filter
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get autocomplete suggestions
      tags:
      - Coins
...
  /value-history:
    get:
      description: Returns value snapshots over time for charting collection value
        trends.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ValueSnapshot'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get value history
      tags:
      - Coins
...
securityDefinitions:
  ApiKeyAuth:
    description: Enter your API key, e.g. "ak_a1b2c3d4..."
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Enter your JWT token with the Bearer prefix, e.g. "Bearer eyJhbGci..."
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"